Added new helper function to check if user is Admin

+++ app/views/posts/show.blade.php
@@ -15,12 +15,10 @@
 
     <hr>
 
-    @if (Auth::check())
-    @if (is_owner_or_admin(Auth::user(), $post))
+    @if (Auth::check() && is_owner_or_admin(Auth::user(), $post))
         <a class="btn btn-xs btn-info" href="{{ $post->edit_url }}"><i class="fa fa-edit"></i> Edit</a>
         <button class="btn btn-xs btn-warning pull-right" data-toggle="modal" data-target="#delete"><i class="fa fa-trash-o"></i> Delete</button>
         {{ HTML::deleteModal('delete','posts','Post', $post->id) }}
-    @endif
     @endif
 
     {{ HTML::br('2') }}
